Make the property geometryInfo public;
